myproject/requirements.txt
sqlparse>=0.2.4
asgiref==3.7.2
cffi==1.16.0
cryptography==42.0.4
Django==3.2.10
django-cors-headers==4.3.1
djangorestframework==3.14.0
djongo==1.3.6
dnspython==2.6.1
google-re2==1.1
httplib2==0.20.4
lightrun==1.26.0
psutil==5.9.8
pycparser==2.21
PyJWT==2.8.0
pymongo==3.12.1
pyparsing==3.1.1
pytz==2024.1
PyYAML==6.0.1
six==1.16.0
typing-extensions==4.9.0
